[fix](inverted index)Change index_id from int32 to int64 to avoid overflow #35206 (#35573)

+++ be/src/olap/rowset/segment_v2/inverted_index_desc.cpp
@@ -28,13 +28,13 @@ const std::string index_suffix = ".idx";
 const std::string index_name_separator = "_";
 
 std::string InvertedIndexDescriptor::get_temporary_index_path(const std::string& segment_path,
-                                                              uint32_t uuid) {
+                                                              uint64_t uuid) {
     return StripSuffixString(segment_path, segment_suffix) + index_name_separator +
            std::to_string(uuid);
 }
 
 std::string InvertedIndexDescriptor::get_index_file_name(const std::string& segment_path,
-                                                         uint32_t uuid) {
+                                                         uint64_t uuid) {
     return StripSuffixString(segment_path, segment_suffix) + index_name_separator +
            std::to_string(uuid) + index_suffix;
 }
